requirements.txt
flask
openai
requests
gunicorn
langchain
langchain-core
langchain-community
langchain_openai
langchainhub
langchain_text_splitters 
pypdf 
PyPDF2 
python-docx
faiss-cpu
unstructured
fastapi
uvicorn
python-dotenv